+++ src/bin/bat/clap_app.rs
@@ -60,6 +60,22 @@ pub fn build_app(interactive_output: bool) -> Command {
                 ),
         )
         .arg(
+            Arg::new("nonprintable-notation")
+                .long("nonprintable-notation")
+                .action(ArgAction::Set)
+                .default_value("unicode")
+                .value_parser(["unicode", "caret"])
+                .value_name("notation")
+                .hide_default_value(true)
+                .help("Set notation for non-printable characters.")
+                .long_help(
+                    "Set notation for non-printable characters.\n\n\
+                    Possible values:\n  \
+                    * unicode (␇, ␊, ␀, ..)\n  \
+                    * caret   (^G, ^J, ^@, ..)",
+                ),
+        )
+        .arg(
             Arg::new("plain")
                 .overrides_with("plain")
                 .overrides_with("number")