+++ src/nonprintable_notation.rs
@@ -1,0 +1,12 @@
+/// How to print non-printable characters with
+/// [crate::config::Config::show_nonprintable]
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
+#[non_exhaustive]
+pub enum NonprintableNotation {
+    /// Use caret notation (^G, ^J, ^@, ..)
+    Caret,
+
+    /// Use unicode notation (␇, ␊, ␀, ..)
+    #[default]
+    Unicode,
+}