+++ tests/integration_tests.rs
@@ -1624,6 +1624,17 @@ fn show_all_extends_tab_markers_to_next_tabstop_width_8() {
 }
 
 #[test]
+fn show_all_with_caret_notation() {
+    bat()
+        .arg("--show-all")
+        .arg("--nonprintable-notation=caret")
+        .arg("nonprintable.txt")
+        .assert()
+        .stdout("hello·world^J\n├──┤^M^@^G^H^[")
+        .stderr("");
+}
+
+#[test]
 fn no_paging_arg() {
     bat()
         .arg("--no-paging")